Include command line argument for data file path Add command line print for packet errors Refactor project structure for improved organisation
