+++ test_input_sensor_processor.py
@@ -1,0 +1,10 @@
+import pytest
+import input_sensor_processor
+
+def test_my_first_one():
+
+    print("\n I'm inside my first test!")
+
+def test_my_second_one():
+
+    print("\n I'm inside my second test!")
